Migrated Common Components to UI Shared lib
