--- apps/sp-student/src/components/atoms/button/button.module.scss
@@ -1,16 +1,0 @@
-.btn {
-  border: none;
-  color: white;
-  background-color: #6521FF;
-  font-size: 14px;
-  padding: 8px 10px;
-  border-radius: 48px;
-  cursor: pointer;
-  &:active {
-    box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 12px;
-  }
-  &:hover{
-    background-color: #ffffff;
-    border: 1px solid #6521FF;
-  }
-}
