--- apps/sp-student/src/components/atoms/button/index.js
@@ -1,2 +1,0 @@
-import Button from "./Button";
-export default Button;
