--- apps/sp-student/src/components/atoms/card/index.js
@@ -1,1 +1,0 @@
-export { default } from './Card';
