--- apps/sp-student/src/components/atoms/icon/Icon.js
@@ -1,23 +1,0 @@
-import {
-  ClockCircleFilled,
-  CommentOutlined,
-  DashboardOutlined,
-  EnvironmentOutlined,
-  LogoutOutlined,
-  ProfileOutlined,
-  KeyOutlined,
-  SettingOutlined,
-  UserOutlined,
-} from '@ant-design/icons';
-
-export {
-  ClockCircleFilled,
-  CommentOutlined,
-  DashboardOutlined,
-  EnvironmentOutlined,
-  LogoutOutlined,
-  ProfileOutlined,
-  KeyOutlined,
-  SettingOutlined,
-  UserOutlined,
-};
