--- apps/sp-student/src/components/atoms/icon/index.js
@@ -1,1 +1,0 @@
-export * from "./Icon";
