--- apps/sp-student/src/components/atoms/input/index.js
@@ -1,3 +1,0 @@
-import Input from './text';
-import Textarea from './textarea';
-export { Input, Textarea };
