--- apps/sp-student/src/components/atoms/menu/Menu.js
@@ -1,3 +1,0 @@
-import {  Menu } from 'antd';
-
-export default Menu;
