--- apps/sp-student/src/components/atoms/menu/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Menu";
