--- apps/sp-student/src/components/atoms/progress/index.js
@@ -1,1 +1,0 @@
-export { default } from './Progress';
