--- apps/sp-student/src/components/atoms/typography/index.js
@@ -1,3 +1,0 @@
-import Title from "./Title";
-import Label from "./Label";
-export { Label, Title }
