+++ libs/ui-shared/src/lib/components/atoms/avatar/Avatar.js
