+++ libs/ui-shared/src/lib/components/atoms/avatar/index.js
