+++ libs/ui-shared/src/lib/components/atoms/button/Button.js
@@ -1,21 +1,21 @@
-import React from "react";
-import PropTypes from "prop-types";
+import React from 'react';
+import PropTypes from 'prop-types';
 
 // Components
-import { Button as AntdButton } from "antd";
+import { Button as AntdButton } from 'antd';
 
 // Helpers
-import { getButtonClassName } from "./helpers/button.general";
+import { getButtonClassName } from './helpers/button.general';
 
 // Constants
-import BUTTON_TYPES from "./constants/button.types";
-import BUTTON_SIZES from "./constants/button.sizes";
+import BUTTON_TYPES from './constants/button.types';
+import BUTTON_SIZES from './constants/button.sizes';
 
 const Button = (props) => {
-  const { className, label, type, htmlType, children, ...restProps } = props;
+  const { className, label, type, children, ...restProps } = props;
   const buttonClassName = getButtonClassName(className, type);
   return (
-    <AntdButton {...restProps} type={type} htmlType={htmlType} className={buttonClassName}>
+    <AntdButton {...restProps} type={type} className={buttonClassName}>
       {label || children}
     </AntdButton>
   );
@@ -25,14 +25,12 @@ Button.propTypes = {
   className: PropTypes.string,
   label: PropTypes.string,
   type: PropTypes.string,
-  htmlType: PropTypes.string,
   size: PropTypes.string,
 };
 
 Button.defaultProps = {
   className: undefined,
   label: undefined,
-  htmlType: "string",
   type: BUTTON_TYPES.PRIMARY,
   size: BUTTON_SIZES.MEDIUM,
 };
