+++ libs/ui-shared/src/lib/components/atoms/button/button.module.scss
@@ -1,8 +1,6 @@
-.button {
-}
+.button {}
 
-.primaryButton {
-}
+.primaryButton {}
 
 .tertiaryButton {
   background-color: #f6f6f6 !important;
