+++ libs/ui-shared/src/lib/components/atoms/button/constants/button.sizes.js
@@ -1,6 +1,6 @@
-const SMALL = "small";
-const MEDIUM = "middle";
-const LARGE = "large";
+const SMALL = 'small';
+const MEDIUM = 'middle';
+const LARGE = 'large';
 
 const SIZES = {
   SMALL,
