+++ libs/ui-shared/src/lib/components/atoms/button/constants/button.types.js
@@ -1,10 +1,10 @@
-const PRIMARY = "submit";
-const GHOST = "ghost";
-const DASHED = "dashed";
-const LINK = "link";
-const TEXT = "text";
-const DEFAULT = "default";
-const TERTIARY = "tertiary";
+const PRIMARY = 'primary';
+const GHOST = 'ghost';
+const DASHED = 'dashed';
+const LINK = 'link';
+const TEXT = 'text';
+const DEFAULT = 'default';
+const TERTIARY = 'tertiary';
 
 const BUTTON_TYPES = {
   PRIMARY,
