+++ libs/ui-shared/src/lib/components/atoms/button/index.js
@@ -1,5 +1,5 @@
-import BUTTON_TYPES from "./constants/button.types";
-import BUTTON_SIZES from "./constants/button.sizes";
+import BUTTON_TYPES from './constants/button.types';
+import BUTTON_SIZES from './constants/button.sizes';
 
-export { default } from "./Button";
+export { default } from './Button';
 export { BUTTON_TYPES, BUTTON_SIZES };
