+++ libs/ui-shared/src/lib/components/atoms/card/card.module.scss
