+++ libs/ui-shared/src/lib/components/atoms/card/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Card";
+export { default } from './Card';
