+++ libs/ui-shared/src/lib/components/atoms/icon/Icon.js
@@ -1,61 +1,23 @@
 import {
-  CloudUploadOutlined,
-  FacebookFilled,
-  InstagramOutlined,
-  LeftOutlined,
-  RightOutlined,
-  StarFilled,
-  StarTwoTone,
-  TwitterOutlined,
-  VideoCameraFilled,
-  StarOutlined,
-  EditOutlined,
-  DeleteOutlined,
-  DownOutlined,
+  ClockCircleFilled,
+  CommentOutlined,
+  DashboardOutlined,
+  EnvironmentOutlined,
   LogoutOutlined,
-  HomeOutlined,
-  SearchOutlined,
-  TeamOutlined,
-  HomeFilled,
-  BellOutlined,
-  BellFilled,
-  HeartOutlined,
-  HeartFilled,
-  CaretRightOutlined,
-  PlusOutlined,
-  MinusOutlined,
-  PlusCircleOutlined,
-  EyeOutlined,
-  InfoCircleOutlined,
-} from "@ant-design/icons";
+  ProfileOutlined,
+  KeyOutlined,
+  SettingOutlined,
+  UserOutlined,
+} from '@ant-design/icons';
 
 export {
-  DownOutlined,
-  EditOutlined,
-  DeleteOutlined,
-  CloudUploadOutlined,
-  FacebookFilled,
-  InstagramOutlined,
-  LeftOutlined,
-  RightOutlined,
-  StarFilled,
-  StarOutlined,
-  StarTwoTone,
-  TwitterOutlined,
-  VideoCameraFilled,
+  ClockCircleFilled,
+  CommentOutlined,
+  DashboardOutlined,
+  EnvironmentOutlined,
   LogoutOutlined,
-  HomeOutlined,
-  SearchOutlined,
-  TeamOutlined,
-  HomeFilled,
-  BellOutlined,
-  BellFilled,
-  HeartOutlined,
-  HeartFilled,
-  CaretRightOutlined,
-  PlusOutlined,
-  MinusOutlined,
-  PlusCircleOutlined,
-  EyeOutlined,
-  InfoCircleOutlined,
+  ProfileOutlined,
+  KeyOutlined,
+  SettingOutlined,
+  UserOutlined,
 };
