+++ libs/ui-shared/src/lib/components/atoms/icon/index.js
@@ -1,30 +1,1 @@
-export {
-  CloudUploadOutlined,
-  FacebookFilled,
-  InstagramOutlined,
-  LeftOutlined,
-  RightOutlined,
-  StarFilled,
-  StarTwoTone,
-  TwitterOutlined,
-  VideoCameraFilled,
-  StarOutlined,
-  EditOutlined,
-  DeleteOutlined,
-  DownOutlined,
-  LogoutOutlined,
-  HomeOutlined,
-  SearchOutlined,
-  TeamOutlined,
-  HomeFilled,
-  BellOutlined,
-  BellFilled,
-  HeartOutlined,
-  HeartFilled,
-  CaretRightOutlined,
-  PlusOutlined,
-  MinusOutlined,
-  PlusCircleOutlined,
-  EyeOutlined,
-  InfoCircleOutlined,
-} from "./Icon";
+export * from './Icon';
