--- libs/ui-shared/src/lib/components/atoms/input/constants/input.sizes.js
@@ -1,11 +1,0 @@
-const SMALL = "small";
-const MEDIUM = "default";
-const LARGE = "large";
-
-const SIZES = {
-  SMALL,
-  MEDIUM,
-  LARGE,
-};
-
-export default SIZES;
