--- libs/ui-shared/src/lib/components/atoms/input/constants/input.types.js
@@ -1,9 +1,0 @@
-const TEXT = "text";
-const TEXTAREA = "textarea";
-
-const TYPES = {
-  TEXT,
-  TEXTAREA,
-};
-
-export default TYPES;
