+++ libs/ui-shared/src/lib/components/atoms/input/index.js
@@ -1,5 +1,3 @@
-import INPUT_SIZES from "./constants/input.sizes";
-import INPUT_TYPES from "./constants/input.types";
-
-export { default } from "./Input";
-export { INPUT_SIZES, INPUT_TYPES };
+import Input from './text';
+import Textarea from './textarea';
+export { Input, Textarea };
