+++ libs/ui-shared/src/lib/components/atoms/logo/Logo.js
