+++ libs/ui-shared/src/lib/components/atoms/logo/index.js
