+++ libs/ui-shared/src/lib/components/atoms/logo/logo.module.scss
