+++ libs/ui-shared/src/lib/components/atoms/progress/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Progress";
+export { default } from './Progress';
